create and add NavigationBar component
